Esercizio 4 sistemato

+++ Alessio.Exercises4/AbstactClass/Asset.cs
@@ -7,13 +7,13 @@ namespace Alessio.Exercises4
 {
     public abstract class Asset
     {
-        int     _valueAsset;
-        string  _name;
+        decimal     _valueAsset;
+        string      _name;
 
-        public int ValueAsset { get { return _valueAsset; } set { _valueAsset =value; } } 
+        public decimal ValueAsset { get { return _valueAsset; } set { _valueAsset =value; } } 
         public string Name { get { return _name; } set {  _name = value; } }
 
-        public Asset(string name, int valueAsset) 
+        public Asset(string name, decimal valueAsset) 
         {
             Name = name;
             ValueAsset = valueAsset;
